// hub/frontend/lib/api/generated/model/registryServerDetailResponse.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Wardn Hub API
 * OpenAPI spec version: 0.2.30
 */
import type { RegistryServerRead } from './registryServerRead';
import type { RegistryServerVersionRead } from './registryServerVersionRead';

export interface RegistryServerDetailResponse {
  server: RegistryServerRead;
  versions?: RegistryServerVersionRead[];
}